Fix lapasAddUser in user mode
If using lapasAddUser in user-mode before, any read on /etc/passwd and /etc/shadow
from then on result in "stale file handle" because the overlay's lowerdir has changed
on the server. We now use the given administration password to remount the root
filesystem, so we flush the cashes.
This even directly makes the new user available *puts sunglasses on*