Finally get rid of uid 1000 hack
Finally was able to remove the hack that all users have the same uid.
We now use bindfs as overlayfs lowerdir to map the uid from base user lapas (1000)
to the uid of the current user on login.
Now that every user has their own uid, a lot of bugs disappeared
- The systemd usersession properly starts
- Unlocking the screen now works (before needed lapas password to unlock)
- XFCE config daemon properly starts
- And many more